Fixed some bugs in files.xsd schema (some unneeded references to ObservableList and some needed weren't there).
When starting the viewer, each XML file that gets fully loaded is checked,
and if it isn't a valid model, it is removed from the category list.

Fixed a bug in JAXBOrigamiHandler.loadModel(): when the XML file didn't
either contain a right root element ("origami"), it didn't throw an
appropriate exception, but it ended with using a previously loaded model as
the result, which is wrong.