Changed the way FoldLines are given to StepRenderer - now each line has its own LineArray - this is due to the fact that we need to highlight single lines - and that's done via Appearance object assigned to a Geometry - and LineArray is the best suitable Geometry.
The new algorithm tries to pass each line only once (even if it's part
of many folds), so no multiple lines should be detected in the model.

The new algorithm, is, however, slow - it runs in O(n^2), where n is
the number of FoldLines (many optimizations are one, but the
asymptotic complexity is still O(n^2)). Practically the algorithm will
run quite fast.