A large reimplementation of the event model on generated classes.
A hierarchical event system has been established to allow the
programmer setup listeners which will "survive" changes in mutable
objects. The programmer should add the listener to the topmost object
(Origami or Listing) and use the hierarchical property name to achieve
this. Observable lists in inner mutable objects can be listened to
this way, too.

Also fixed a bunch of logical and listener-related bugs mainly in
Editor GUI and OrigamiPropertiesFrame.