Implemented new idea how to perform bendPaper - now it should really respect the given direction of the fold. The idea is based on comparing the layers' normals with the current screen normal.
Got rid of specifying second layers for double folds
(thunderbolt and crimp) - the layers that were rotated by the first
fold operation are automatically used as the layers for the second
fold to operate on.

Also got rid of specifying a refPoint for thunderbolt fold - now it is
autocomputed from the second fold line.

Still has some bugs.